refactor: RankingService 구조 개편 및 Test Profile 분리
- RankingService를 Interface로 전환하고 구현체 분리
- 'test' Profile용 NoopRankingService 도입
- Controller 및 Scheduler에 Profile Annotation 적용
- RankingController의 Parameter Handling 및 Null Check 개선
