+++ src/main/java/com/eventitta/common/config/redis/RedisConfig.java
@@ -20,7 +20,7 @@ import org.springframework.data.redis.serializer.StringRedisSerializer;
 @Slf4j
 @Configuration
 @EnableCaching
-@Profile("prod")
+@Profile("!test")
 public class RedisConfig {
 
     @Value("${spring.data.redis.host}")
