+++ src/main/java/com/eventitta/gamification/service/RankingSyncService.java
@@ -7,6 +7,7 @@ import com.eventitta.user.domain.User;
 import com.eventitta.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -24,6 +25,7 @@ import java.util.stream.Collectors;
  */
 @Slf4j
 @Service
+@Profile("!test")  // 테스트 프로파일에서 비활성화
 @RequiredArgsConstructor
 public class RankingSyncService {
